Implement kirigami-2 with input nodes
Complete input node implementation for Kirigami 2.
The main change is the modification of assertions which fold over nodes.
For such assertions, we replace the existing function with one that
skips nodes that are input nodes.

This implementation adds input branches for init, trans and merge, using
the standard form we've used in the past. This has a downside in that
these branches typically increase the size of the expressions, sometimes
to the point that they seem to be less performant than the nodeless
implementation. More optimization may be possible, however.